overlord/servicestate: mv ensureSnapServicesForGroup to new file
This new file will be the home of the new handlers we execute as tasks for the
change based quota group modifications.

Also slightly refactor to use an options struct instead of defaulting to empty
things
